Object.m (compare): Add type id.
2008-06-10  Kai Tietz  <kai.tietz@onevision.com>

	* Object.m (compare): Add type id.
	* objc/Object.h: Likewise.
	* archive.c (objc_read_class): Use size_t to extend version to be
	size of pointer scalar width.
	* sendmsg.c (rtx): Undefine it before redefinition.
	(__objc_print_dtable_stats): Cast arguments to long as intended.

From-SVN: r136617

+++ libobjc/archive.c
@@ -908,7 +908,7 @@ objc_read_class (struct objc_typed_stream *stream, Class *class)
 
 	  objc_read_unsigned_long (stream, &version);
 	  objc_hash_add (&stream->class_table,
-			 (*class)->name, (void *)version);
+			 (*class)->name, (void *) ((size_t) version));
 	}
 
       else if ((buf[0]&_B_CODE) == _B_UCOMM)
